+++ pkg/util/limiter/query_limiter.go
@@ -18,6 +18,7 @@ var (
 	ctxKey                    = &queryLimiterCtxKey{}
 	ErrMaxSeriesHit           = "the query hit the max number of series limit (limit: %d series)"
 	ErrMaxChunkBytesHit       = "the query hit the aggregated chunks size limit (limit: %d bytes)"
+	ErrMaxDataBytesHit        = "the query hit the aggregated data size limit (limit: %d bytes)"
 	ErrMaxChunksPerQueryLimit = "the query hit the max number of chunks limit (limit: %d chunks)"
 )
 
@@ -26,16 +27,18 @@ type QueryLimiter struct {
 	uniqueSeries   map[model.Fingerprint]struct{}
 
 	chunkBytesCount atomic.Int64
+	dataBytesCount  atomic.Int64
 	chunkCount      atomic.Int64
 
 	maxSeriesPerQuery     int
 	maxChunkBytesPerQuery int
+	maxDataBytesPerQuery  int
 	maxChunksPerQuery     int
 }
 
 // NewQueryLimiter makes a new per-query limiter. Each query limiter
 // is configured using the `maxSeriesPerQuery` limit.
-func NewQueryLimiter(maxSeriesPerQuery, maxChunkBytesPerQuery int, maxChunksPerQuery int) *QueryLimiter {
+func NewQueryLimiter(maxSeriesPerQuery, maxChunkBytesPerQuery, maxChunksPerQuery, maxDataBytesPerQuery int) *QueryLimiter {
 	return &QueryLimiter{
 		uniqueSeriesMx: sync.Mutex{},
 		uniqueSeries:   map[model.Fingerprint]struct{}{},
@@ -43,6 +46,7 @@ func NewQueryLimiter(maxSeriesPerQuery, maxChunkBytesPerQuery int, maxChunksPerQ
 		maxSeriesPerQuery:     maxSeriesPerQuery,
 		maxChunkBytesPerQuery: maxChunkBytesPerQuery,
 		maxChunksPerQuery:     maxChunksPerQuery,
+		maxDataBytesPerQuery:  maxDataBytesPerQuery,
 	}
 }
 
@@ -56,7 +60,7 @@ func QueryLimiterFromContextWithFallback(ctx context.Context) *QueryLimiter {
 	ql, ok := ctx.Value(ctxKey).(*QueryLimiter)
 	if !ok {
 		// If there's no limiter return a new unlimited limiter as a fallback
-		ql = NewQueryLimiter(0, 0, 0)
+		ql = NewQueryLimiter(0, 0, 0, 0)
 	}
 	return ql
 }
@@ -94,6 +98,17 @@ func (ql *QueryLimiter) AddChunkBytes(chunkSizeInBytes int) error {
 	}
 	if ql.chunkBytesCount.Add(int64(chunkSizeInBytes)) > int64(ql.maxChunkBytesPerQuery) {
 		return fmt.Errorf(ErrMaxChunkBytesHit, ql.maxChunkBytesPerQuery)
+	}
+	return nil
+}
+
+// AddDataBytes adds the queried data bytes and returns an error if the limit is reached.
+func (ql *QueryLimiter) AddDataBytes(dataSizeInBytes int) error {
+	if ql.maxDataBytesPerQuery == 0 {
+		return nil
+	}
+	if ql.dataBytesCount.Add(int64(dataSizeInBytes)) > int64(ql.maxDataBytesPerQuery) {
+		return fmt.Errorf(ErrMaxDataBytesHit, ql.maxDataBytesPerQuery)
 	}
 	return nil
 }